Emit buddy_joined for others on joining
An activity will often maintain a list of buddies.

For a leader, it is easy to maintain the list, after sharing the
activity, by receiving the buddy_joined and buddy_left signals.

For a non-leader, it was not easy to maintain the list, after joining a
shared activity, without calling underneath CollabWrapper into sugar3.

When an activity has been joined, iterate through the buddies and emit
buddy_joined for each.

Also rewrite documentation accordingly, and simplify.

Backport from sugarlabs/collabwrapper:84be150