Handle network loss of buddy
Buddy leaving is normally handled by receipt of a message from the buddy
that is leaving.  This does not happen when a buddy leaves in an
uncontrolled fashion, such as power failure, suspend, force quit, or
wireless network disconnection.  The only hope is that Telepathy notices
and CollabWrapper sends a buddy-left signal.